requirements.txt
aiogram==3.0.0b7
mysql-connector-python==8.0.33
python-dotenv==1.0.0